Technical Report 2
This report is ongoing work and focuses on the TLS handshake; the miTLS
handshake module is used as a precise specification. This allows us to prove the
handshake as it is used on the internet. In particular we investigate new and
unfortunately rather strong agile assumptions; assumptions in which the same
cryptographic keying material is used by different cryptographic primitives.
This is not a drawback of our approach but a consequence of bad protocol design
choices made in the past. Our model is compatible with the results in Report 2
and thus enhances our understanding of the cryptographic assumptions implicitly
made in the design of TLS.

While the first report focuses on the close relation between our analysis and
the miTLS implementation and is from a formal methods perspective, the second is
directed more at cryptographers, both works however share the twin ambition of
making a cryptographic statement about an executable artifact, a reference
implementation that describes TLS in less ambiguous terms than the various RFCs.
This more precise treatment led to the detection of numerous corner cases. This
in turn led to new attacks described in both works.